fixing AsLazyValue tests
